pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "plagiarism_detector"
version = "2.0.1"
description = "Advanced Multi-Agent Internet Plagiarism Detection"
authors = [{name = "M.Tech Student", email = "student@example.com"}]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    "langchain>=0.2.16,<0.3.0",
    "langchain-openai>=0.1.14,<0.2.0",
    "langchain-community>=0.2.16,<0.3.0",
    "langgraph>=0.2.20,<0.3.0",
    "fastapi>=0.115.0,<0.116.0",
    "uvicorn[standard]>=0.30.6,<0.31.0",
    "pydantic>=2.9.2,<3.0.0",
    "pypdf>=4.3.1,<5.0.0",
    "faiss-cpu>=1.8.0,<2.0.0",
    "python-multipart>=0.0.9,<0.0.10",
    "python-dotenv>=1.0.1,<1.1.0",
    "httpx>=0.27.2,<0.28.0",
    "streamlit>=1.32.0,<1.33.0",
    "duckduckgo-search>=6.2.11,<7.0.0",
    "tavily-python>=0.4.0,<0.5.0",
    "numpy>=1.26.4,<1.27.0",
    "requests>=2.32.3,<2.33.0",
    "beautifulsoup4>=4.12.3,<4.13.0",
    "networkx>=3.3,<3.4",
    "arxiv>=2.1.3,<2.2.0",
    "langchain-groq>=0.1.10",
    "datetime>=6.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.3,<9.0.0",
    "pytest-asyncio>=0.23.6,<0.24.0",
    "pytest-cov>=5.0.0,<6.0.0",
    "black>=24.4.2,<24.5.0",
    "ruff>=0.4.4,<0.5.0",
    "mypy>=1.10.0,<1.11.0"
]

# 🔥 FIXED: Correct Wheel Builder Config
[tool.hatch.build.targets.wheel]
packages = ["src"]
sources = ["src"]

# ✅ FIXED: Simple entry points (no complex args)
[project.scripts]
start-api = "plagiarism_detector.app:start_api"
start-dashboard = "plagiarism_detector.dashboard:start_dashboard"
test-api = "plagiarism_detector.tests.test_api:test_complete_plagiarism_workflow"

# Dev tools
[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
